fix(layout): consolidate root html tags and use server-side redirect
