+++ src/app/[locale]/layout.tsx
@@ -1,10 +1,8 @@
 import type { Metadata } from "next";
-import { Inter, Noto_Sans_Devanagari } from "next/font/google";
 import { NextIntlClientProvider } from 'next-intl';
 import { notFound } from 'next/navigation';
 import { routing } from "@/i18n/routing";
 import { generateLocaleMetadata } from "@/i18n/metadata";
-import "./globals.css";
 
 import enMessages from "@/messages/en.json";
 import hiMessages from "@/messages/hi.json";
@@ -17,18 +15,6 @@ const messagesMap: Record<string, Record<string, unknown>> = {
   en: enMessages as Record<string, unknown>,
   hi: hiMessages as Record<string, unknown>,
 };
-
-const inter = Inter({
-  subsets: ["latin"],
-  variable: "--font-inter",
-  display: "swap",
-});
-
-const devanagari = Noto_Sans_Devanagari({
-  subsets: ["devanagari"],
-  variable: "--font-devanagari",
-  display: "swap",
-});
 
 export function generateStaticParams() {
   return routing.locales.map((locale) => ({
@@ -45,7 +31,7 @@ export async function generateMetadata({
   return generateLocaleMetadata(locale, 'metadata');
 }
 
-export default async function RootLayout({
+export default async function LocaleLayout({
   children,
   params
 }: Readonly<{
@@ -61,24 +47,20 @@ export default async function RootLayout({
   const messages = messagesMap[locale] || enMessages;
 
   return (
-    <html lang={locale} className={`${inter.variable} ${devanagari.variable}`} suppressHydrationWarning>
-      <body className="antialiased font-sans bg-background text-foreground">
-        <NextIntlClientProvider locale={locale} messages={messages} formats={{}} timeZone="UTC" now={new Date(0)}>
-          <ThemeProvider
-            attribute="class"
-            defaultTheme="system"
-            enableSystem
-            disableTransitionOnChange
-          >
-            <Navbar />
-            <LenisProvider>
-              <main className="pt-20">
-                {children}
-              </main>
-            </LenisProvider>
-          </ThemeProvider>
-        </NextIntlClientProvider>
-      </body>
-    </html>
+    <NextIntlClientProvider locale={locale} messages={messages} formats={{}} timeZone="UTC" now={new Date(0)}>
+      <ThemeProvider
+        attribute="class"
+        defaultTheme="system"
+        enableSystem
+        disableTransitionOnChange
+      >
+        <Navbar />
+        <LenisProvider>
+          <main className="pt-20">
+            {children}
+          </main>
+        </LenisProvider>
+      </ThemeProvider>
+    </NextIntlClientProvider>
   );
 }
