+++ src/app/layout.tsx
@@ -1,9 +1,31 @@
 import { ReactNode } from 'react';
+import { Inter, Noto_Sans_Devanagari } from "next/font/google";
+import "./[locale]/globals.css";
+
+const inter = Inter({
+  subsets: ["latin"],
+  variable: "--font-inter",
+  display: "swap",
+});
+
+const devanagari = Noto_Sans_Devanagari({
+  subsets: ["devanagari"],
+  variable: "--font-devanagari",
+  display: "swap",
+});
 
 type Props = {
   children: ReactNode;
 };
 
 export default function RootLayout({ children }: Props) {
-  return children;
+  return (
+    <html className={`${inter.variable} ${devanagari.variable}`} suppressHydrationWarning>
+      <body className="antialiased font-sans bg-background text-foreground">
+        {children}
+      </body>
+    </html>
+  );
 }
+
+
